fix: create new focusChange action
