+++ web/src/lib/components/shared-components/search-bar/search-bar.svelte
@@ -12,6 +12,7 @@
   import { getMetadataSearchQuery } from '$lib/utils/metadata-search';
   import { handlePromiseError } from '$lib/utils';
   import { shortcut } from '$lib/utils/shortcut';
+  import { focusChange } from '$lib/utils/focus-change';
 
   export let value = '';
   export let grayTheme: boolean;
@@ -94,7 +95,7 @@
   }}
 />
 
-<div class="w-full relative" use:clickOutside={{ onOutclick: onFocusOut, onFocusOut }}>
+<div class="w-full relative" use:clickOutside={{ onOutclick: onFocusOut }} use:focusChange={{ onFocusOut }}>
   <form
     draggable="false"
     autocomplete="off"
